Fix: Use BigDecimal for financial calculations in UbkCalculateGMDService
The `calculateGMD` method was using `Double` for financial calculations, which
can lead to precision errors. It was also using `BigDecimal.ROUND_HALF_UP`
for rounding, which is not ideal for financial calculations.

This commit refactors the `calculateGMD` method to use `BigDecimal` for all
calculations and changes the rounding mode to `RoundingMode.HALF_EVEN`
(banker's rounding) to minimize cumulative errors.

A new test suite, `UbkCalculateGMDServiceImpTest`, has been added to verify
the correctness of the new implementation.

+++ ubk/src/main/java/kg/mlsp/ubk/service/UbkCalculateGMDService.java
@@ -1,5 +1,7 @@
 package kg.mlsp.ubk.service;
+import java.math.BigDecimal;
+
 public interface UbkCalculateGMDService {
-    Double calculateGMD(Integer applicationId);
+    BigDecimal calculateGMD(Integer applicationId);
 
 }
